edit message and icon

+++ clients/admin-ui/src/features/datastore-connections/system_portal_config/ConnectionForm.tsx
@@ -8,13 +8,12 @@ import {
   ChakraBox as Box,
   ChakraFlex as Flex,
   ChakraStack as Stack,
-  Tooltip,
+  Modal,
   useChakraDisclosure as useDisclosure,
 } from "fidesui";
 import React, { useEffect, useState } from "react";
 
 import { useAppSelector } from "~/app/hooks";
-import ConfirmationModal from "~/features/common/modals/ConfirmationModal";
 import { useDeleteConnectorTemplateMutation } from "~/features/connector-templates/connector-template.slice";
 import ConnectorTemplateUploadModal from "~/features/connector-templates/ConnectorTemplateUploadModal";
 import { ConnectorParameters } from "~/features/datastore-connections/system_portal_config/forms/ConnectorParameters";
@@ -26,7 +25,6 @@ import {
   SystemType,
 } from "~/types/api";
 
-import { CUSTOM_INTEGRATION_INDICATOR } from "../constants";
 import {
   selectDatastoreConnectionFilters,
   useGetAllDatastoreConnectionsQuery,
@@ -79,21 +77,35 @@ const ConnectionForm = ({ connectionConfig, systemFidesKey }: Props) => {
   }, [data]);
 
   const uploadTemplateModal = useDisclosure();
-  const deleteTemplateModal = useDisclosure();
-  const [deleteConnectorTemplate, { isLoading: isDeleting }] =
-    useDeleteConnectorTemplateMutation();
+  const [modalApi, modalContext] = Modal.useModal();
+  const [deleteConnectorTemplate] = useDeleteConnectorTemplateMutation();
 
-  const handleDeleteCustomIntegration = async () => {
-    if (selectedConnectionOption?.identifier) {
-      try {
-        await deleteConnectorTemplate(
-          selectedConnectionOption.identifier,
-        ).unwrap();
-        deleteTemplateModal.onClose();
-      } catch {
-        // Error handling is managed by RTK Query
-      }
-    }
+  const handleRemoveCustomIntegration = () => {
+    modalApi.confirm({
+      title: "Remove",
+      type: "error",
+      content: (
+        <>
+          This will remove the custom integration template and update all
+          systems and connections that use it. All instances will revert to the
+          Fides-provided default integration template.
+          <br />
+          <br />
+          This change applies globally and cannot be undone. Are you sure you
+          want to proceed?
+        </>
+      ),
+      okText: "Remove",
+      okButtonProps: { danger: true },
+      centered: true,
+      onOk: async () => {
+        if (selectedConnectionOption?.identifier) {
+          await deleteConnectorTemplate(
+            selectedConnectionOption.identifier,
+          ).unwrap();
+        }
+      },
+    });
   };
 
   /* STEPS TO UNIFY the database and saas forms
@@ -112,13 +124,6 @@ const ConnectionForm = ({ connectionConfig, systemFidesKey }: Props) => {
             onChange={setSelectedConnectionOption}
             disabled={Boolean(connectionConfig && connectionConfig !== null)}
           />
-          {selectedConnectionOption?.custom && (
-            <Tooltip title="Custom integration" placement="top">
-              <Box as="span" cursor="pointer" fontSize="lg">
-                {CUSTOM_INTEGRATION_INDICATOR}
-              </Box>
-            </Tooltip>
-          )}
           {!connectionConfig && orphanedConnectionConfigs.length > 0 ? (
             <OrphanedConnectionModal
               connectionConfigs={orphanedConnectionConfigs}
@@ -127,6 +132,19 @@ const ConnectionForm = ({ connectionConfig, systemFidesKey }: Props) => {
           ) : null}
 
           <Restrict scopes={[ScopeRegistryEnum.CONNECTOR_TEMPLATE_REGISTER]}>
+            {connectionConfig &&
+              selectedConnectionOption?.custom &&
+              selectedConnectionOption?.default_connector_available && (
+                <Button
+                  type="link"
+                  danger
+                  data-testid="delete-custom-integration-btn"
+                  onClick={handleRemoveCustomIntegration}
+                  className="ml-2"
+                >
+                  Remove
+                </Button>
+              )}
             <Button
               htmlType="submit"
               data-testid="upload-btn"
@@ -135,17 +153,6 @@ const ConnectionForm = ({ connectionConfig, systemFidesKey }: Props) => {
             >
               Upload integration
             </Button>
-            {connectionConfig &&
-              selectedConnectionOption?.custom &&
-              selectedConnectionOption?.default_connector_available && (
-                <Button
-                  data-testid="delete-custom-integration-btn"
-                  onClick={deleteTemplateModal.onOpen}
-                  className="ml-2"
-                >
-                  Use Fides provided template
-                </Button>
-              )}
           </Restrict>
         </Stack>
 
@@ -153,26 +160,7 @@ const ConnectionForm = ({ connectionConfig, systemFidesKey }: Props) => {
           isOpen={uploadTemplateModal.isOpen}
           onClose={uploadTemplateModal.onClose}
         />
-        <ConfirmationModal
-          isOpen={deleteTemplateModal.isOpen}
-          onClose={deleteTemplateModal.onClose}
-          onConfirm={handleDeleteCustomIntegration}
-          title="Use Fides provided template"
-          message={
-            <>
-              This will remove the custom integration template and update all
-              connections that use it by falling back to the Fides provided
-              template.
-              <br />
-              <br />
-              Are you sure you want to proceed?
-            </>
-          }
-          cancelButtonText="No"
-          continueButtonText="Yes"
-          isLoading={isDeleting}
-          testId="delete-custom-integration-modal"
-        />
+        {modalContext}
       </Flex>
       {selectedConnectionOption?.type &&
       [
